Refactor focus tracking
