+++ Wrecept.Wpf/App.xaml.cs
@@ -114,7 +114,6 @@ public static IServiceProvider Provider => Services ?? throw new InvalidOperatio
         services.AddSingleton<StatusBarViewModel>();
         services.AddSingleton<INotificationService, MessageBoxNotificationService>();
         services.AddSingleton<ScreenModeManager>();
-        services.AddSingleton<IFocusTrackerService, FocusTrackerService>();
         services.AddSingleton<KeyboardManager>();
         services.AddSingleton<FocusManager>();
         services.AddTransient<ProgressViewModel>();
