--- Wrecept.Wpf/Services/IFocusTrackerService.cs
@@ -1,10 +1,0 @@
-using System.Windows;
-using System.Windows.Input;
-
-namespace Wrecept.Wpf.Services;
-
-public interface IFocusTrackerService
-{
-    void Update(string viewKey, IInputElement element);
-    IInputElement? GetLast(string viewKey);
-}
