+++ Wrecept.Wpf/ViewModels/InvoiceCreatePromptViewModel.cs
@@ -25,9 +25,8 @@ public partial class InvoiceCreatePromptViewModel : ObservableObject
     {
         await _parent.CreateInvoiceAsync(Number);
         _parent.InlinePrompt = null;
-        var tracker = App.Provider.GetRequiredService<IFocusTrackerService>();
         var focus = App.Provider.GetRequiredService<FocusManager>();
-        var last = tracker.GetLast("InvoiceEditorView");
+        var last = focus.GetLast("InvoiceEditorView");
         focus.RequestFocus(last);
     }
 
@@ -35,9 +34,8 @@ public partial class InvoiceCreatePromptViewModel : ObservableObject
     private void Cancel()
     {
         _parent.InlinePrompt = null;
-        var tracker = App.Provider.GetRequiredService<IFocusTrackerService>();
         var focus = App.Provider.GetRequiredService<FocusManager>();
-        var last = tracker.GetLast("InvoiceEditorView");
+        var last = focus.GetLast("InvoiceEditorView");
         focus.RequestFocus(last);
     }
 }
