+++ docs/progress/2025-07-03_21-34-45_code_agent.md
@@ -1,0 +1,1 @@
+- FocusManager és FocusTrackerService összevonva, hivatkozások frissítve.
